fix: P0 protocol-sync correctness — v6 createTask arity + voting ConfigKey map
task create / create-batch called the REMOVED 7-arg createTask — hard revert
(missing selector) on every TaskManager v6 org:
- task create: new --deadline (ISO/unix/relative) + --completion-window flags,
  9-arg createTask on v6 orgs; pre-v6 orgs auto-fall back to a legacy
  7-arg fragment interface (feature-detected via implementation bytecode);
  deadline flags on a legacy org exit 4 with an actionable message
- task create-batch: single-tx createTasksBatch on v6 (all-or-nothing,
  per-row deadline/completionWindow in JSONL + batch-wide flag defaults),
  legacy per-task loop preserved for pre-v6 orgs

vote propose-config silently set THRESHOLD when asked for quorum (key 0 is
THRESHOLD on both voting contracts):
- correct v6 key map: threshold H0/DD0 uint8, quorum H3/DD4 uint32 (voter
  COUNT since PR #119 — semantics note printed), executor H2/DD1,
  target-allowed DD-only (deprecated no-op on Hybrid now errors), hat-allowed
  DD3; propose-quorum now encodes uint32
- vote create: hat IDs parsed as BigNumber (parseInt lost precision above
  2^53 — real Hats IDs are uint256)

29 new handler tests pin the encoded calldata (key numbers, value types,
tuple order) so this cannot silently regress.

+++ src/commands/vote/propose-quorum.ts
@@ -9,12 +9,13 @@ import * as output from '../../lib/output';
 import { resolveVotingContracts } from './helpers';
 
 /**
- * setConfig keys for quorum:
- *   HybridVoting: key 3
- *   DirectDemocracyVoting: key 4
+ * setConfig keys for quorum (contracts origin/main, v6):
+ *   HybridVoting: key 3 (QUORUM)
+ *   DirectDemocracyVoting: key 4 (QUORUM)
  *
- * Discovered by reverse-engineering Proposal #0 (setQuorum to 1).
- * The voting contracts use setConfig(uint8, bytes) instead of a direct setter.
+ * The voting contracts use setConfig(uint8, bytes) instead of a direct
+ * setter, and decode the value as uint32. Since PR #119 quorum is a
+ * minimum voter COUNT (0 disables), not a percentage.
  */
 const HYBRID_QUORUM_KEY = 3;
 const DD_QUORUM_KEY = 4;
@@ -49,7 +50,7 @@ export const proposeQuorumHandler = {
 
       // Encode setConfig calls for both voting contracts
       const iface = new ethers.utils.Interface(['function setConfig(uint8 key, bytes value)']);
-      const encodedValue = ethers.utils.defaultAbiCoder.encode(['uint256'], [newQuorum]);
+      const encodedValue = ethers.utils.defaultAbiCoder.encode(['uint32'], [newQuorum]);
 
       const hybridCall = iface.encodeFunctionData('setConfig', [HYBRID_QUORUM_KEY, encodedValue]);
